fix: initial
The Z parameter of the playing card size is set to 1.
Shadow Depth Bias increased to prevent texture jitter.
Fixed a shop bug when there are 10 cards in hand.
Improved display of cards in hand.
Minor changes to objects on the main scene.
Reduced the number of "magic numbers" in the code.